test reader class
